wip: add 'Order Completed' for mixpanel actions
